build-windows: Fix broken output zip due to incorrect line endings
There is a possibility that any file be altered incorrectly in Windows, resulting in a file with CRLF endings.

This is not the behaviour we want, as the whole thing is going to be fed on to a Linux-based system; we want a CR ending instead.

Using dos2unix, recursively change the endings for all file under the temp-magisk folder. It's a bit of an overkill, but it should handle the problem finely.

* dos2unix binary is also added along in this commit.

* Amended: Make sure busybox is copied after the line ending conversion. Not doing so will break the binary, as it is not meant to be processed by dos2unix.